added edit button to infowindow, added markerEditor component/css
I added an edit button to the infowindow. I also added an markerEditor component. The goal is to slide this markerEditor modal component onto the screen when the user clicks the edit button. A big problem now is that click events aren't registering inside the info window. help me zac!! :D